refactor(cli): extract install runtime profile and release helpers
